fix(cm-qa): resolve TS case-sensitivity — merge questionCard into QuestionCard
Lowercase questionCard.test.tsx caused TS1149 on Linux CI (case-insensitive
macOS hid the conflict locally). Consolidated new deacon-qbl tests into the
canonical PascalCase filename and removed the duplicate.
